requirements.txt
crewai
crewai-tools
langchain-community
langchain-ollama
python-dotenv
pykrx
pandas